requirements.txt
# export PATH="/usr/local/cuda-11.8/bin:$PATH"
timm==0.9.1
Pillow==10.0.0
blobfile==2.1.1
mypy==1.7.0
numpy==1.23.5
requests==2.28.1
einops==0.7.0
tensorboardX==2.6.2.2
deepspeed==0.12.3
scipy==1.10.0
seldon-core==1.17.1
python-dotenv
boto3
shapely
opencv-python
pandas

#Optional<dependency by alignment model for now>
fastapi
uvicorn
python-multipart
#required by ocr
pymupdf
paddlepaddle
paddleclas
faiss-cpu